Removed chat logic

+++ include/header.h
@@ -19,9 +19,9 @@
 
 typedef struct date
 {
-    size_t day;
-    size_t month; 
-    size_t year;
+    int day;
+    int month; 
+    int year;
 }date_t;
 
 // Message types
